Replace the Hello World image with a rails 7 app.
This deploys a working Rails 7 app, which uses
a postgres database and a redis instance to
support hotwire.

Currently, this deployment uses the RDS and
elasticache instances from my deployment of the
same app. on Heroku.

This is a starting point which will let me replace
first the RDS instance and then the redis instance
with ones managed by terraform.